requirements.txt
Flask==2.2.5
Markdown==3.4.4
MarkupSafe==2.1.5
Pillow==9.5.0
PyWavelets==1.3.0
Werkzeug==2.2.3
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cycler==0.11.0
dnspython==2.3.0
eventlet==0.38.2
flatbuffers==25.2.10
fonttools==4.38.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.62.3
h5py==3.8.0
idna==3.10
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.6
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
libclang==18.1.1
matplotlib==3.5.3
netifaces==0.10.6
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
pip==24.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-engineio==3.8.2.post1
python-socketio==4.2.1
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools>=41.0.0
shapely==2.0.7
six==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tqdm==4.67.1
typing-extensions==4.7.1
urllib3==2.0.7
wheel==0.42.0
wrapt==1.16.0
zipp==3.15.0
